Add `At` draw functions taking transform matrices
Much of the view transformation logic in the skill tree can be replaced
with a per-frame transformation matrix going from tree space to screen
space. By separating the geometry from its view the pieces of data are
likely to be the same for the whole frame or even for the life of the
program.

This could be further improved by caching static geometry in the runtime
so that only a handle is needed to draw it with a transform.