Update connectors to draw prepared meshes
With the introduction of prepared 2D meshes there is some significant
runtime performance improvements enabled by not sending the same data
over and over across the language boundary.

This change builds a mesh up front for connectors when the passive tree
is constructed and draws them with the tree transform whenever they are
processed.